requirements.txt
elevenlabs
moviepy
scipy
google-generativeai
openai-whisper
pyttsx3
python-dotenv